Move demo to full
